Migrate my Modal.tsx code to typescript and apply some rules and refactoring
